Refactored code again, putting figures in view utils folders
