--- webviz_subsurface/plugins/_well_analysis/_figures/__init__.py
@@ -1,2 +1,0 @@
-from .well_control_figure import create_well_control_figure
-from .well_overview_figure import WellOverviewFigure, format_well_overview_figure
