+++ webviz_subsurface/plugins/_well_analysis/_views/__init__.py
@@ -1,6 +1,10 @@
-from ._well_control import (
+from ._well_control_view import (
     WellControlPressurePlotOptions,
     WellControlSettings,
     WellControlView,
 )
-from ._well_overview import WellOverviewFilters, WellOverviewSettings, WellOverviewView
+from ._well_overview_view import (
+    WellOverviewFilters,
+    WellOverviewSettings,
+    WellOverviewView,
+)
