+++ webviz_subsurface/plugins/_well_analysis/_views/_well_control_view/__init__.py
@@ -1,0 +1,1 @@
+from ._view import WellControlPressurePlotOptions, WellControlSettings, WellControlView
