+++ webviz_subsurface/plugins/_well_analysis/_views/_well_control_view/_utils/__init__.py
@@ -1,0 +1,1 @@
+from ._well_control_figure import create_well_control_figure
