+++ webviz_subsurface/plugins/_well_analysis/_views/_well_control_view/_utils/_well_control_figure.py
@@ -6,8 +6,8 @@ import plotly.graph_objects as go
 from plotly.subplots import make_subplots
 from webviz_config import WebvizConfigTheme
 
-from ...._utils.colors import StandardColors
-from .._types import NodeType, PressurePlotMode
+from ......_utils.colors import StandardColors
+from ...._types import NodeType, PressurePlotMode
 
 
 def create_well_control_figure(
