+++ webviz_subsurface/plugins/_well_analysis/_views/_well_overview_view/__init__.py
@@ -1,0 +1,1 @@
+from ._view import WellOverviewFilters, WellOverviewSettings, WellOverviewView
