+++ webviz_subsurface/plugins/_well_analysis/_views/_well_overview_view/_utils/__init__.py
@@ -1,0 +1,1 @@
+from ._well_overview_figure import WellOverviewFigure, format_well_overview_figure
