+++ webviz_subsurface/plugins/_well_analysis/_views/_well_overview_view/_utils/_well_overview_figure.py
@@ -8,8 +8,8 @@ import plotly.graph_objects as go
 from plotly.subplots import make_subplots
 from webviz_config import WebvizConfigTheme
 
-from .._types import ChartType
-from .._utils import EnsembleWellAnalysisData
+from ...._types import ChartType
+from ...._utils import EnsembleWellAnalysisData
 
 
 class WellOverviewFigure:
