+++ webviz_subsurface/plugins/_well_analysis/_views/_well_overview_view/_view.py
@@ -8,10 +8,10 @@ from dash.development.base_component import Component
 from webviz_config import WebvizConfigTheme
 from webviz_config.webviz_plugin_subclasses import SettingsGroupABC, ViewABC
 
-from .._figures import WellOverviewFigure, format_well_overview_figure
-from .._plugin_ids import PluginIds
-from .._types import ChartType
-from .._utils import EnsembleWellAnalysisData
+from ..._plugin_ids import PluginIds
+from ..._types import ChartType
+from ..._utils import EnsembleWellAnalysisData
+from ._utils import WellOverviewFigure, format_well_overview_figure
 
 
 class WellOverviewSettings(SettingsGroupABC):
